slight refactor, add test mod
