+++ soar-db/src/lib.rs
@@ -1,28 +1,23 @@
 // (S)ubnets (O)ptimistic (A)daptive (R)eplay DB
+extern crate clarity;
 extern crate stacks_common;
 
 use std::collections::HashMap;
 
+use crate::memory::MemoryBackingStore;
+use clarity::vm::database::ClarityBackingStore;
 use stacks_common::types::chainstate::StacksBlockId;
+
+pub mod memory;
+
+#[cfg(test)]
+pub mod tests;
 
 pub trait SoarBackingStore {}
 
 /// Key-Value Store with edit log
 pub struct SoarDB {
     storage: MemoryBackingStore,
-}
-
-pub struct BlockData {
-    put_log: Vec<PutCommand>,
-    parent: Option<StacksBlockId>,
-    height: u64,
-    id: StacksBlockId,
-}
-
-pub struct MemoryBackingStore {
-    current_block: Option<StacksBlockId>,
-    entries: HashMap<String, String>,
-    blocks: HashMap<StacksBlockId, BlockData>,
 }
 
 #[derive(Clone)]
@@ -35,6 +30,12 @@ pub struct PutCommand {
 }
 
 impl SoarDB {
+    pub fn new_memory() -> SoarDB {
+        SoarDB {
+            storage: MemoryBackingStore::new(),
+        }
+    }
+
     /// If the DB has a block, then the current block should be returned
     /// If there is *no* block data yet, this will return none
     pub fn current_block(&self) -> Option<&StacksBlockId> {
@@ -54,6 +55,25 @@ impl SoarDB {
         let parent_ht = self.storage.get_block_height(&parent)?;
         assert_eq!(block_ht - 1, parent_ht);
         Ok((parent, parent_ht))
+    }
+
+    pub fn add_genesis(
+        &mut self,
+        block: StacksBlockId,
+        put_list: Vec<PutCommand>,
+    ) -> Result<(), String> {
+        if !self.storage.is_empty()? {
+            return Err("Attempted to add genesis block to DB that already has data".into());
+        }
+
+        self.storage.store_genesis_block(block.clone(), put_list);
+        for put in put_list.into_iter() {
+            self.storage.apply_put(put);
+        }
+
+        self.storage.set_current_block(block);
+
+        Ok(())
     }
 
     pub fn add_block_ops(
@@ -159,114 +179,5 @@ impl SoarDB {
             self.storage.set_current_block(block);
         }
         Ok(())
-    }
-}
-
-impl MemoryBackingStore {
-    pub fn has_block(&self, block: &StacksBlockId) -> bool {
-        self.blocks.contains_key(block)
-    }
-
-    pub fn reapply_block(&mut self, block: &StacksBlockId) -> Result<(), String> {
-        let block_data = self.blocks.get(block).ok_or_else(|| "No such block")?;
-
-        for command in block_data.put_log.clone().into_iter() {
-            self.apply_put(command);
-        }
-
-        Ok(())
-    }
-
-    pub fn undo_block(&mut self, expected_cur_block: &StacksBlockId) -> Result<(), String> {
-        if self.current_block.is_none() || self.current_block.as_ref() != Some(expected_cur_block) {
-            return Err("Expected current block does not match storage's view of current block during rollback".into());
-        }
-
-        let block_data = self
-            .blocks
-            .get(expected_cur_block)
-            .expect("Could not find block data for current block");
-        let parent = block_data.parent.clone();
-
-        // undo each operation in reverse order from the edit log
-        for put_command in block_data.put_log.clone().into_iter().rev() {
-            self.undo_put(put_command);
-        }
-
-        // operations are undone, now set the current_block to the parent
-        self.current_block = parent;
-
-        Ok(())
-    }
-
-    pub fn get_block_parent(&self, block: &StacksBlockId) -> Result<StacksBlockId, String> {
-        match self.blocks.get(&block) {
-            Some(data) => match data.parent.as_ref() {
-                Some(parent) => Ok(parent.clone()),
-                None => Err("Block is zero-height and has no parent".into()),
-            },
-            None => Err(format!("{} not found in storage", block)),
-        }
-    }
-
-    pub fn get_block_height(&self, block: &StacksBlockId) -> Result<u64, String> {
-        match self.blocks.get(&block) {
-            Some(data) => Ok(data.height),
-            None => Err(format!("{} not found in storage", block)),
-        }
-    }
-
-    pub fn set_current_block(&mut self, block: StacksBlockId) {
-        self.current_block = Some(block);
-    }
-
-    pub fn current_block(&self) -> Option<&StacksBlockId> {
-        self.current_block.as_ref()
-    }
-
-    pub fn store_block_data(
-        &mut self,
-        block: StacksBlockId,
-        parent: StacksBlockId,
-        put_log: Vec<PutCommand>,
-    ) -> Result<(), String> {
-        let parent_height = match self.blocks.get(&parent) {
-            Some(parent_data) => Ok(parent_data.height),
-            None => Err("Parent block has not been processed yet"),
-        }?;
-
-        let prior = self.blocks.insert(
-            block.clone(),
-            BlockData {
-                id: block,
-                parent: Some(parent),
-                put_log,
-                height: parent_height
-                    .checked_add(1)
-                    .ok_or_else(|| "Block height overflowed u64")?,
-            },
-        );
-        assert!(
-            prior.is_none(),
-            "Stored block data over an existing block entry"
-        );
-        Ok(())
-    }
-
-    pub fn apply_put(&mut self, command: PutCommand) {
-        self.entries.insert(command.key, command.value);
-    }
-
-    pub fn undo_put(&mut self, command: PutCommand) {
-        let old_value = if let Some(old_value) = command.prior_value {
-            self.entries.insert(command.key, old_value)
-        } else {
-            self.entries.remove(&command.key)
-        };
-        assert_eq!(
-            old_value,
-            Some(command.value),
-            "Undo operation applied to an entry that had an unexpected value"
-        );
     }
 }
