+++ soar-db/src/tests.rs
@@ -1,0 +1,6 @@
+use crate::SoarDB;
+
+#[test]
+fn simple_storage_chain() {
+    let db = SoarDB::new_memory();
+}
